split validation/error function into separate page
